html5demos.com online / offline event fix.
The property on document.body is called ononline and onoffline
(lowercase second 'o').

Also, according the html5 spec, the ononline IDL attributes on the Window
object, body element and frameset elements all link to the event handler
implemented by the Window object. This means that setting the IDL attribute
on one of these will automatically set the value on the others.